+++ histogram/src/sliding_window/atomic.rs
@@ -1,6 +1,6 @@
 use super::*;
 use crate::atomic::Histogram as AtomicHistogram;
-use core::sync::atomic::AtomicU64;
+use core::sync::atomic::AtomicU32;
 
 /// A type of histogram that reports on the distribution of values across a
 /// moving window of time. For example, the distribution of values for the past
@@ -86,7 +86,7 @@ impl Histogram {
     /// This is useful if you need access to the raw bucket counts or if you are
     /// planning to update from some external source that uses the same
     /// bucketing strategy. Be sure to use with `snapshot()`.
-    pub fn as_slice(&self) -> &[AtomicU64] {
+    pub fn as_slice(&self) -> &[AtomicU32] {
         self.live.as_slice()
     }
 
@@ -104,7 +104,7 @@ impl Histogram {
     /// This is a convenience method that uses `Instant::now()` as the time
     /// associated with the observation. If you already have a timestamp, you
     /// may wish to use the `add_at` instead.
-    pub fn add(&self, value: u64, count: u64) -> Result<(), Error> {
+    pub fn add(&self, value: u64, count: u32) -> Result<(), Error> {
         self.add_at(Instant::now(), value, count)
     }
 
@@ -136,7 +136,7 @@ impl Histogram {
     /// If the instant is within the window, the increment will be attributed to
     /// the most recent time slide regardless of the true position within the
     /// sliding window.
-    pub fn add_at(&self, instant: Instant, value: u64, count: u64) -> Result<(), Error> {
+    pub fn add_at(&self, instant: Instant, value: u64, count: u32) -> Result<(), Error> {
         self.tick_to(instant);
 
         self.live.add(value, count)
